2021/6/9 19:52 提交

+++ mall-seckill/src/main/java/com/learn/mall/seckill/vo/SecKillSkuVo.java
@@ -39,7 +39,7 @@ public class SecKillSkuVo {
     /**
      * 每人限购数量
      */
-    private BigDecimal seckillLimit;
+    private Integer seckillLimit;
     /**
      * 排序
      */
